Change VersionAndShareCount layout to [Version:48][ShareCount:16]
Increase version bits from 32 to 48 for better ABA problem resistance.
Reduce share count to 16 bits (max 65535 concurrent borrows is sufficient).

+++ src/Paradise.ECS/Memory/ChunkHandle.cs
@@ -5,8 +5,8 @@ namespace Paradise.ECS;
 /// Prevents direct unsafe usage of Chunk pointers/structs.
 /// </summary>
 /// <param name="Id">The index of the chunk in the ChunkManager.</param>
-/// <param name="Version">Incrementing version for stale handle detection (wraps on overflow).</param>
-internal readonly record struct ChunkHandle(int Id, uint Version)
+/// <param name="Version">Incrementing version for stale handle detection (48 bits, wraps on overflow).</param>
+internal readonly record struct ChunkHandle(int Id, ulong Version)
 {
     /// <summary>
     /// The Invalid handle.
